#11 2-06 도메인별로 분류하기
question, answer, user 도메인 패키지 생성 후 분류에 맞게 이동
