+++ src/test/java/com/mysite/sbb/SbbApplicationTests.java
@@ -1,5 +1,9 @@
 package com.mysite.sbb;
 
+import com.mysite.sbb.answer.Answer;
+import com.mysite.sbb.answer.AnswerRepository;
+import com.mysite.sbb.question.Question;
+import com.mysite.sbb.question.QuestionRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -19,7 +23,7 @@ class SbbApplicationTests {
 	private QuestionRepository questionRepository;
 
 	@Autowired
-	private AnswerRepository  answerRepository;
+	private AnswerRepository answerRepository;
 
 	@Test
 	void testJpa() {
